% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_shrink_spd.R
\name{cov_shrink_spd}
\alias{cov_shrink_spd}
\title{Asymptotic Stein-type shrinkage of SPD covariance estimate.}
\usage{
cov_shrink_spd(x)
}
\arguments{
\item{x}{an \eqn{nxp} matrix with \eqn{n-}observations of a \eqn{p-}vector}
}
\value{
Dense covariance matrix that is SPD.
}
\description{
\code{cov_shrink_spd} is the covariance estimator in \pkg{GraphSPME}.
}
\details{
The standard frequentist covariance matrix estimate is calculated.
Using techniques of Touloumis (2015) the matrix is shrunk towards a 
sparse target matrix, set to be the diagonal matrix of the frequentist
estimate. The amount of shrinkage is found using asymptotic results
found in the reference paper.
}
\seealso{
\code{\link{prec_sparse}}
}
