% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphspme-package.R
\docType{package}
\name{GraphSPME}
\alias{GraphSPME}
\title{Graphical Sparse Precision Matrix Estimation}
\description{
\code{GraphSPME} is a package for sparse precision matrix estimation through
a graph.
It employs the method of Lunde 2022, imported through the \code{graph_spme} function.
}
\details{
Important functions:

\itemize{
\item \code{\link{prec_sparse}}: Sparse precision matrix estimation through a graph
\item \code{\link{cov_shrink_spd}}: Asymptotic Stein-type shrinkage of SPD covariance estimate
}

See individual function documentation for usage.
}
\author{
Berent Ånund Strømnes Lunde
}
