% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_aic.R
\name{prec_aic}
\alias{prec_aic}
\title{AIC for precision estimation.}
\usage{
prec_aic(x, Prec)
}
\arguments{
\item{x}{an \eqn{nxp} matrix with \eqn{n-}observations of a \eqn{p-}vector}

\item{Prec}{a \eqn{pxp} sparse precision matrix.}
}
\value{
The value of the penalized negative log-likelihood.
}
\description{
\code{prec_aic} gives the AIC-penalized Gaussian negative log-likelihood in \pkg{GraphSPME}.
}
\details{
The AIC-penalized Gaussian negative log-likelihood given by

\deqn{AIC(\Lambda) = \frac{1}{2} tr(S\Lambda)-\log|\Lambda| + c )}{
AIC(\Lambda) = 0.5(tr(S\Lambda)-\log|\Lambda|) + c}
}
\seealso{
\code{\link{prec_nll}}
}
