% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_nll.R
\name{prec_nll}
\alias{prec_nll}
\title{Gaussian negative log-likelihood relevant for precision estimation.}
\usage{
prec_nll(x, Prec)
}
\arguments{
\item{x}{an \eqn{nxp} matrix with \eqn{n-}observations of a \eqn{p-}vector}

\item{Prec}{a \eqn{pxp} sparse precision matrix.}
}
\value{
The value of the negative log-likelihood.
}
\description{
\code{prec_nll} gives the Gaussian negative log-likelihood in \pkg{GraphSPME}.
}
\details{
The Gaussian negative log-likelihood given by

\deqn{l(\Lambda) = \frac{1}{2} tr(S\Lambda)-\log|\Lambda| )}{
l(\Lambda) = 0.5(tr(S\Lambda)-\log|\Lambda|)}
}
\seealso{
\code{\link{prec_aic}}
}
