% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_sparse.R
\name{prec_sparse}
\alias{prec_sparse}
\title{Graphical Sparse Precision Matrix Estimation}
\usage{
prec_sparse(x, Z, shrinkage = TRUE)
}
\arguments{
\item{x}{an \eqn{nxp} matrix with \eqn{n-}observations of a \eqn{p-}vector}

\item{Z}{a sparse matrix indicating non-zero elements of the precision matrix}

\item{shrinkage}{if shrinkage should be applied to block-frequentist-covariance see details for more}
}
\value{
Sparse precision matrix that is SPD.
}
\description{
\code{prec_sparse} is the precision matrix estimator in \pkg{GraphSPME}.
}
\details{
The algorithm utilizes the knowledge of the sparsity provided through
\code{Z}, using the algorithm of Le (2021). If \code{shrinkage=TRUE} then
the method of Le (2021) is modified by the method in 
Lunde (2022) to iteratively employ asymptotic shrinkage when building
the sparse precision matrix estimator.

The algorithm should be robust in the face of extremely high dimensions.
}
\seealso{
\code{\link{cov_shrink_spd}}
}
