% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_sparse.R
\name{prec_sparse}
\alias{prec_sparse}
\title{Graphical Sparse Precision Matrix Estimation}
\usage{
prec_sparse(
  X,
  Graph,
  markov_order = 1,
  shrinkage = TRUE,
  symmetrization = TRUE
)
}
\arguments{
\item{X}{an \eqn{nxp} matrix with \eqn{n-}observations of a \eqn{p-}vector}

\item{Graph}{a \eqn{pxp} sparse matrix representing a graph with vertices
corresponding to columns of \code{X} and edges to direct connections
(neighborhood) between columns of \code{X}}

\item{markov_order}{non-negative int representing the markov order of data
\code{X} wrt \code{Graph}}

\item{shrinkage}{if shrinkage should be applied to
block-frequentist-covariance see details for more}

\item{symmetrization}{if symmetry of the precision estimate should be
ensured through symmetry conversion. See details for more}
}
\value{
Sparse precision matrix that is SPD.
}
\description{
\code{prec_sparse} is the precision matrix estimator in \pkg{GraphSPME}.
}
\details{
The algorithm utilizes the knowledge of the graph provided through
\code{Graph} and \code{markov_order} to create sparse matrix \code{Z} indicating the
non-zero elements of the precision matrix of \code{X}.

The resulting sparse matrix \code{Z} is then utilized to estimate a sparse
precision matrix using the algorithm of Le (2021).

If \code{shrinkage=TRUE} then
the method of Le (2021) is modified by the method in
Lunde (2022) to iteratively employ asymptotic shrinkage when building
the sparse precision matrix estimator.

Employing shrinkage ensures numerical stability while improving the resulting
precision/covariance estimate. This is particularly noticeable when the
problem involves many neighbors for each vertex and/or has a high Markov
order.

If \code{symmetrization=TRUE} then the symmetry conversion
\deqn{\hat{\Lambda}=0.5(\tilde{\Lambda}+\tilde{\Lambda}^\top)}{
0.5(\Lambda+\Lambda^T)}
is returned.

The algorithm should be robust in the face of extremely high dimensions.
}
\seealso{
\code{\link{cov_shrink_spd}}
}
